Reorg source for clarity
