+++ src/PrivateKey.js
@@ -8,7 +8,7 @@ import base58 from './base58';
 import base64 from './base64';
 import Network from './Network';
 import PublicKey from './PublicKey';
-import { generatePublicKey } from './secp256k1';
+import { generatePublicKey } from './PrivateKey/secp256k1';
 
 const BYTES_LENGTH: number = 32;
