+++ src/index.js
@@ -1,5 +1,6 @@
 import Address from './Address';
 import Input from './Input';
+import MnemonicSeed from './MnemonicSeed';
 import Network from './Network';
 import OpCode from './OpCode';
 import opCodes from './opCodes';
@@ -12,6 +13,7 @@ import Transaction from './Transaction';
 module.exports = {
   Address,
   Input,
+  MnemonicSeed,
   Network,
   OpCode,
   opCodes,
