+++ src/opCodes/index.js
@@ -1,0 +1,16 @@
+// @flow
+import OpCode from '../OpCode';
+
+import OP_CHECKSIG from './OP_CHECKSIG';
+import OP_DUP from './OP_DUP';
+import OP_EQUALVERIFY from './OP_EQUALVERIFY';
+import OP_HASH160 from './OP_HASH160';
+
+const codes = {
+  OP_CHECKSIG,
+  OP_DUP,
+  OP_EQUALVERIFY,
+  OP_HASH160,
+}
+
+export default codes;
